extract VirtualScroll from ChatArea
